+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 
 ## [14.1.0][14.1.0] - 2016-08-25
 ### Added
+- [Feature] Add TabNav react component
 - [Feature] Add isFilter prop to <Input> for search icon (#488)
 
 ### Changed